refactor: Move browser execution in dedicated script
