+++ .env.example
@@ -3,4 +3,6 @@ LIVEKIT_DOMAIN="example.com"
 LIVEKIT_KEY="API_KEY"
 LIVEKIT_SECRET="API_SECRET"
 LIVEKIT_ROOM="racecast"
+LIVEKIT_IDENTITY="car"
 LOG_LEVEL="info"
+CHROME_PATH="/usr/bin/chromium"
